+++ scanpipe/tests/test_api.py
@@ -315,7 +315,8 @@ class ScanPipeAPITest(TransactionTestCase):
         self.assertEqual(1, len(response.data))
         resource = response.data[0]
         self.assertEqual(
-            ["pkg:deb/debian/adduser@3.118?arch=all"], resource["for_packages"]
+            ["pkg:deb/debian/adduser@3.118?uuid=610bed29-ce39-40e7-92d6-fd8b"],
+            resource["for_packages"],
         )
         self.assertEqual("filename.ext", resource["path"])
 
@@ -628,7 +629,7 @@ class ScanPipeAPITest(TransactionTestCase):
 
     def test_scanpipe_api_serializer_get_serializer_fields(self):
         self.assertEqual(31, len(get_serializer_fields(DiscoveredPackage)))
-        self.assertEqual(27, len(get_serializer_fields(CodebaseResource)))
+        self.assertEqual(28, len(get_serializer_fields(CodebaseResource)))
 
         with self.assertRaises(LookupError):
             get_serializer_fields(None)