+++ CHANGELOG.rst
@@ -1,15 +1,15 @@
 Changelog
 =========
 
-v31.0.0 (next)
----------------
+v31.0.0 (2022-08-25)
+--------------------
 
 - WARNING: Drop support for Python 3.6 and 3.7. Add support for Python 3.10.
-  Upgrade Django to version 4.x series.
+  Upgrade Django to version 4.1 series.
 
-- Upgrade ScanCode-toolkit to version v31.
+- Upgrade ScanCode-toolkit to version 31.0.x.
   See https://github.com/nexB/scancode-toolkit/blob/develop/CHANGELOG.rst for an
-  overview of the changes in v31 compared to v30.
+  overview of the changes in the v31 compared to v30.
 
 - Implement run status auto-refresh using the htmx JavaScript library.
   The statuses of queued and running pipeline are now automatically refreshed
@@ -69,17 +69,45 @@ v31.0.0 (next)
   https://github.com/nexB/scancode.io/issues/164
   https://github.com/nexB/scancode.io/issues/464
 
-- Add the ability to filter by empty and none values providing the "EMPTY" magic value
-  to any filters.
+- Add the ability to filter by empty and none values providing the
+  "EMPTY" magic value to any filters.
   https://github.com/nexB/scancode.io/issues/296
 
 - CodebaseResource.name now contains both the bare file name with extension, as
   opposed to just the bare file name without extension.
-
-  - Using a name stripped from its extension was something that was not used in
-    other AboutCode project or tools.
-
+  Using a name stripped from its extension was something that was not used in
+  other AboutCode project or tools.
   https://github.com/nexB/scancode.io/issues/467
+
+- Export current results as XLSX for resource, packages, and errors list views.
+  https://github.com/nexB/scancode.io/issues/48
+
+- Add support for .tgz extension for input files in Docker pipeline
+  https://github.com/nexB/scancode.io/issues/499
+
+- Add support for resource missing file content in details view.
+  Refine the annotation using the new className instead of type.
+  https://github.com/nexB/scancode.io/issues/495
+
+- Change the worksheet names in XLSX output, using the
+  "PACKAGES", "RESOURCES", "DEPENDENCIES", and "ERRORS" names.
+  https://github.com/nexB/scancode.io/issues/511
+
+- Update application Package scanning step to reflect the updates in
+  scancode-toolkit package scanning.
+
+  - Package data detected from a file are now stored on the
+    CodebaseResource.package_data field.
+  - A second processing step is now done after scanning for Package data, where
+    Package Resources are determined and DiscoveredPackages are created.
+
+  https://github.com/nexB/scancode.io/issues/444
+
+- ``CodebaseResource.for_packages`` now returns a list of
+  ``DiscoveredPackage.package_uid`` or ``DiscoveredPackage.package_url`` if
+  ``DiscoveredPackage.package_uid`` is not present. This is done to reflect the
+  how scancode-toolkit's JSON output returns ``package_uid`` in the
+  ``for_packages`` field for Resources.
 
 v30.2.0 (2021-12-17)
 --------------------